benchmark/qutip/requirements.txt
numpy>=1.21
qutip>=5.1
jax[cuda12]<0.5
qutip-jax>=0.1
pytest-benchmark==5.1.0